Integrate v7 virtualization into StructView
